meta_index, a barebones tool for automatically indexing npm projects
Summary:
This is a (currently very CommonJS/npm/git-specific) tool
for indexing an arbitrary repository and its dependencies.
Given a repository root, it will build a vnames.json file,
an indexing script, and a set of entries describing the various
corpora it discovers.

This depends on D604 and is maximally useful with D599.

Reviewers: fromberger, schroederc, shahms

Reviewed By: shahms

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D606