Add member reference support to Java indexer.
Summary: Make the Java indexer handle member references, and add tests that they're handled properly. Because constructor references are written "Type::new" (even while the internal name for a constructor is always "<init>"), I added fake names for "new"s; there might be a better way to do this, but this was a straightforward one.

Reviewers: schroederc

Reviewed By: schroederc

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D696