Allow Java/C++ extractors to force single-file mode with KYTHE_OUTPUT_FILE and clean up some bzl rules
Summary:
This adds a KYTHE_OUTPUT_FILE environment variable to both the Java and C++ extractors
which will be used to specify the single output file name, rather than using a directory.
This was largely to simplify the verify/extract based tests to avoid needing a subsequent
mv in the wrapper script.

It then builds on this to remove some of the incomplete argument escaping by avoiding it entirely
or allowing the shell to do it directly.

Reviewers: schroederc, zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D827