Re-set the workspace name, use PWD rather than TEST_SRCDIR as appropriate or include workspace name in path
Summary:
This change also increases the required Bazel versio to 0.2.1 as that is when the
workspace-name path changes were incorporated.  Basically, we should be limiting the
number of directory changes as a lot of Bazel is built around sticking you in the appropriate
root and passing relative paths. In particular, for tests, PWD = TEST_SRCDIR/{workspace_name}.

Reviewers: schroederc, zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D830