Don't emit an implicit Object superclass for interface types.
Summary:
It is true that all interfaces are subtypes of Object
(see http://docs.oracle.com/javase/specs/jls/se8/html/jls-4.html#jls-4.10.2),
but an interface does not extend Object and is therefore not
a nominal subtype, as implied by the EXTENDS edge.

(The fact that Kythe emits that edge was carelessness on my
part in 2f10c61be6.  The intention was only to emit for classes,
so that the behavior would be uniform whether the "extends Object"
was implicit or explicit. This is not pertinent to interfaces.)

Test Plan:
Verified that new test breaks without the fix, passes with it.
Verified that all javatests pass.

Reviewers: zarko

Reviewed By: zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1004