Add format tokens for nicer signatures
Summary:
This diff adds format tokens that make it possible to generate
nicer signatures. It changes the Java indexer to produce these
signatures.

The tokens provide a way to turn off parent delimiters (so
no more leading dots in .java.lang.String); a way to annotate
a span of text as important (so you can show only "String"
instead of the whole qualified name); and a way to annotate
a span of text as belonging to a parameter list (so you can
elect to drop parameter lists for function signatures).

The diff also changes the Java indexer to use parameter names
rather than types when generating printable method signatures.

Reviewers: fromberger, #core_team, shahms

Reviewed By: fromberger, #core_team, shahms

Subscribers: shahms, fromberger

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1025