Add bounded/* to the schema doc
Summary:
Add documentation of the bounded/upper and bounded/lower edges to the
schema doc. These are used when a type variable is bounded by other
types. In java, this is done with <? extends String> or <? super
String>.

Also, add a test for for Java generics that uses <? super String> since
all the other bounded tests were all covariant.

Also, add IJwB project setup for java indexer so it is easier for people
to set up and use IntelliJ to hack on the Java indexer.

Test Plan: Updated Generics.java test and added example code to schema.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: zarko

Maniphest Tasks: T179

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1116