Fixup the generic node structure for Java
Summary:
According to the schema and other langauges, there should be no source
range that represents the class node. The class name source range should
be used for the abs node. See the updated tests for examples.

Test Plan: Updated test and updated schema with an example for Java.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Maniphest Tasks: T176

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1129