Don't elide anonymous subclasses of Java enum classes.
Summary:
E.g.,

enum Day {
  SUNDAY { String desc() { return "Bloody"; } },
  MONDAY { String desc() { return "Manic"; } },
  ...
  abstract String desc();
}

In this case SUNDAY, MONDAY etc. are actually instances of an
anonymous subclass of Day.  Note that this is not merely an
implementation detail but defined in the language spec itself
(https://docs.oracle.com/javase/specs/jls/se8/html/jls-8.html#jls-8.9.1):
"The optional class body of an enum constant implicitly defines an anonymous
class declaration (§15.9.5) that extends the immediately enclosing enum type."

Previously this case was special-cased in the SignatureGenerator,
with the result that the anonymous class had the same signature as the
constant instance of it, leading to subtle confusion later on.

I'm not sure why this was special-cased, perhaps the thinking was that this
anonymous class was merely an implementation detail (which, as I mention
above, is not actually the case).  Either way, it seems better to me
not to treat this case specially, and certainly not in a way that causes
signature confusion.

So this change removes the special-casing.

It also fixes an over-eager warning log, that fired unnecessarily
for anonymous classes.

Reviewers: #core_team, salguarnieri

Reviewed By: #core_team, salguarnieri

Subscribers: salguarnieri, zarko

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1277