Java code cleanup
Summary:
Remove the isLocalAnalysis parameter, as it was only ever set to false.
Add default cases for non-exhaustive switch statements.
Add a parameter to the relevant method in JavacOptionsUtils to enable grabbing classpath information from the JavaDetails.  (This is a prequel to the changes needed to address T222.

Test Plan: Tests run manually (bazel 0.4.5 causes problems with the presubmit checks).

Reviewers: salguarnieri

Reviewed By: salguarnieri

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1434