Emit a Diagnostic if trying to emit a node from an unknown source.
Summary:
Previously, a NAME node would be emitted in such cases.
However NAME nodes are no longer part of the Kythe schema, so this
is needlessly confusing.

This slightly simplifies the node building code, and removes
a source of confusion.

Test Plan:
All java tests pass.  A c++ test fails for reasons
  unclear to me and unrelated to this change.

Reviewers: schroederc

Reviewed By: schroederc

Subscribers: schroederc, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D1942