Add command line flags to the JVM classfile indexer.
Summary:
The primary goal is to support sending output to a file instead of stdout.
But since there is already an IndexerConfig base class, whose javadoc claims
that it should be common to all indexers, it made sense to base the
classfile indexer's config off that, and employ its other flags, such as
--help.
--default_metadata_corpus comes along for the ride, but its use is vacuous
since the classfile indexer doesn't currently do anything with metadata.

While adding the --out flag, also added --print_statistics,
as that seems likely to be useful.

While I was in this code, also replaced a hand-crafted, and stale,
usage message with the one generated by JCommander.

Reviewers: schroederc, #core_team

Reviewed By: schroederc, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2168